+++ packages/app-bridge/src/react/useAssetChooser.spec.tsx
@@ -6,6 +6,7 @@ import { afterEach, describe, it } from 'vitest';
 import { cleanup, render } from '@testing-library/react';
 
 import type { Asset } from '../types/Asset';
+import { AssetDummy } from '../tests/AssetDummy';
 import { AppBridgeBlock } from '../AppBridgeBlock';
 import { useAssetChooser } from './useAssetChooser';
 import { withAppBridgeBlockStubs } from '../tests/withAppBridgeBlockStubs';
@@ -33,7 +34,7 @@ const AssetChooserDummy = ({
     );
 };
 
-describe('useReadyForPrint hook', () => {
+describe('useAssetChooser hook', () => {
     afterEach(() => {
         cleanup();
     });
@@ -43,14 +44,34 @@ describe('useReadyForPrint hook', () => {
         const { getByTestId } = render(<BlockWithStubs />);
         const openAssetChooserButton = getByTestId(OPEN_ASSET_CHOOSER_BUTTON_ID) as HTMLButtonElement;
         openAssetChooserButton.click();
-        sinon.assert.calledOnce(appBridge.openAssetChooser);
+        sinon.assert.calledWith(appBridge.dispatch, sinon.match.has('name', 'openAssetChooser'));
     });
 
     it('should close the asset chooser', () => {
         const [BlockWithStubs, appBridge] = withAppBridgeBlockStubs(AssetChooserDummy);
         const { getByTestId } = render(<BlockWithStubs />);
-        const openAssetChooserButton = getByTestId(CLOSE_ASSET_CHOOSER_BUTTON_ID) as HTMLButtonElement;
+        const closeAssetChooserButton = getByTestId(CLOSE_ASSET_CHOOSER_BUTTON_ID) as HTMLButtonElement;
+        closeAssetChooserButton.click();
+        sinon.assert.calledWith(appBridge.dispatch, sinon.match.has('name', 'closeAssetChooser'));
+    });
+
+    it('should call the onAssetChosen callback when an asset is chosen', () => {
+        const [BlockWithStubs] = withAppBridgeBlockStubs(AssetChooserDummy);
+        const onAssetChosen = sinon.spy();
+        const { getByTestId } = render(<BlockWithStubs onAssetChosen={onAssetChosen} />);
+        const openAssetChooserButton = getByTestId(OPEN_ASSET_CHOOSER_BUTTON_ID) as HTMLButtonElement;
         openAssetChooserButton.click();
-        sinon.assert.calledOnce(appBridge.closeAssetChooser);
+        sinon.assert.calledWith(onAssetChosen, [AssetDummy.with(123)]);
+    });
+
+    it('should unsubscribe if asset chooser gets opened and closed', () => {
+        const unsubscribeSpy = sinon.spy();
+        const [BlockWithStubs] = withAppBridgeBlockStubs(AssetChooserDummy, { unsubscribe: unsubscribeSpy });
+        const { getByTestId } = render(<BlockWithStubs />);
+        const openTemplateChooserButton = getByTestId(OPEN_ASSET_CHOOSER_BUTTON_ID) as HTMLButtonElement;
+        openTemplateChooserButton.click();
+        const closeTemplateChooserButton = getByTestId(CLOSE_ASSET_CHOOSER_BUTTON_ID) as HTMLButtonElement;
+        closeTemplateChooserButton.click();
+        sinon.assert.calledOnce(unsubscribeSpy);
     });
 });